update all the services

+++ src/Applications/use_case/DeleteCommentUseCase.js
@@ -1,21 +1,14 @@
 class DeleteCommentUseCase {
-  constructor({
-    userRepository,
-    threadRepository,
-    commentRepository,
-    authenticationTokenManager,
-  }) {
+  constructor({ userRepository, threadRepository, commentRepository }) {
     this._userRepository = userRepository;
     this._threadRepository = threadRepository;
     this._commentRepository = commentRepository;
-    this._authenticationTokenManager = authenticationTokenManager;
   }
 
   async execute(id, useCasePayload) {
-    const { username, id: user_id } =
-      await this._authenticationTokenManager.decodePayload(id);
-    await this._userRepository.getUserById(user_id);
+    await this._userRepository.getUserById(id);
     await this._threadRepository.getThreadById(useCasePayload.thread_id);
+    await this._commentRepository.verifyCommentOwner(useCasePayload.id, id);
 
     return this._commentRepository.deleteComment(useCasePayload.id);
   }
